requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0
langchain==0.3.12
langchain-openai==0.2.14
langchain-community==0.3.12
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.3
